--- src/test/java/pages/PlaylistPage.java
@@ -1,10 +1,0 @@
-package pages;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-
-public class PlaylistPage extends BasePage {
-    public PlaylistPage(WebDriver givenDriver) {
-        super(givenDriver);
-    }
-}
