+++ src/test/resources/Regression.xml
@@ -3,12 +3,15 @@
 
 <suite name="Regression Tests" parallel="methods" thread-count="1">
 
+    <parameter name='BaseURL' value="https://bbb.testpro.io/"/>
+
     <test name="Example Test" preserve-order="false">
         <classes>
+            <class name="RegistrationTests"/>
             <class name="LoginTests"/>
             <class name="ProfileTests"/>
             <class name="AllSongsTests"/>
-            <class name="PlaylistTests"/>
+            <class name="HomeTests"/>
         </classes>
     </test>
 </suite>
