+++ src/test/resources/Smoke.xml
@@ -3,6 +3,8 @@
 
 <suite name="Smoke Tests" parallel="methods" thread-count="1">
 
+    <parameter name='BaseURL' value="https://bbb.testpro.io/"/>
+
     <test name="Example Test" preserve-order="false">
         <classes>
             <class name="LoginTests">
@@ -20,14 +22,13 @@
                     <include name="playSong"/>
                 </methods>
             </class>
-            <class name="PlaylistTests">
+            <class name="HomeTests">
                 <methods>
                     <include name="addSongToPlaylist"/>
                     <include name="deletePlaylist"/>
                     <include name="renamePlaylist"/>
                 </methods>
             </class>
-
         </classes>
     </test>
 </suite>
