+++ src/test/resources/TestNG.xml
@@ -1,12 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="TestNG">
+<suite name="TestNG" parallel="methods" thread-count="1">
+    <parameter name='BaseURL' value="https://bbb.testpro.io/"/>
 
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="LoginTests"/>
-            <class name="BaseTest"/>
+            <class name="HomeTests">
+                <methods>
+                    <include name="searchSong"/>
+                </methods>
+            </class>
         </classes>
     </test>
 </suite>
